desacoplamento BookExercise
